+++ source/skills.cpp
@@ -1146,7 +1146,7 @@ void cSkills::Mine(int s)
 			for( gridY = 1; gridY < 410; gridY++ )
 	 		{
 				oreamount[gridX][gridY] = resource.ore;
-				oretime[gridX][gridY] = uiCurrentTime + resource.oretime * CLOCKS_PER_SEC;
+				oretime[gridX][gridY] = (unsigned int) (uiCurrentTime + resource.oretime * CLOCKS_PER_SEC);
 			}
 		}
 	}
@@ -1255,7 +1255,7 @@ void cSkills::Mine(int s)
 			else
 				break;
 		}
-		oretime[oreX][oreY] = uiCurrentTime + resource.oretime * CLOCKS_PER_SEC;	// reset ore regen timer
+		oretime[oreX][oreY] = (unsigned int) (uiCurrentTime + resource.oretime * CLOCKS_PER_SEC);	// reset ore regen timer
 	}
 	
 	if( oreamount[oreX][oreY] > resource.ore )	// if there's too much ore, then put a ceiling on it
@@ -1996,8 +1996,8 @@ void cSkills::TreeTarget(int s)
 		{
 			for(b=1;b<410;b++)
 			{
-				logamount[a][b]=resource.logs;
-				logtime[a][b]=curtime+resource.logtime*CLOCKS_PER_SEC;
+				logamount[a][b] = resource.logs;
+				logtime[a][b] = (unsigned int) (curtime + resource.logtime * CLOCKS_PER_SEC);
 			}
 		}
 		printf(" Done.\n");
@@ -2035,7 +2035,7 @@ void cSkills::TreeTarget(int s)
 				logamount[a][b]++;
 			else break;
 		}
-		logtime[a][b]=curtime+resource.logtime*CLOCKS_PER_SEC;//10 more mins
+		logtime[a][b] = (unsigned int) (curtime + resource.logtime * CLOCKS_PER_SEC);//10 more mins
 	}
 	
 	if(logamount[a][b]>resource.logs) logamount[a][b]=resource.logs;
@@ -3484,7 +3484,7 @@ void cSkills::SpiritSpeak(int s)  // spirit speak time, on a base of 30 seconds 
 	
 	//chars[currchar[s]].spiritspeaktimer=spiritspeak_data.spiritspeaktimer+chars[currchar[s]].skill[SPIRITSPEAK]/50+chars[currchar[s]].in; // spirit speak duration
 	//modified by AntiChrist
-	chars[currchar[s]].spiritspeaktimer = uiCurrentTime + CLOCKS_PER_SEC * ( spiritspeak_data.spiritspeaktimer + chars[currchar[s]].skill[SPIRITSPEAK] / 10 + chars[currchar[s]].in ); // spirit speak duration
+	chars[currchar[s]].spiritspeaktimer = (unsigned int) (uiCurrentTime + CLOCKS_PER_SEC * ( spiritspeak_data.spiritspeaktimer + chars[currchar[s]].skill[SPIRITSPEAK] / 10 + chars[currchar[s]].in)); // spirit speak duration
 	//	chars[currchar[s]].spiritspeaktimer = uiCurrentTime + CLOCKS_PER_SEC * 5;
 	// care to make up your mind on which to use??? I chose the more complex one (Abaddon)
 }
@@ -4104,7 +4104,7 @@ void cSkills::SkillUse(int s, int x) // Skill is clicked on the skill list
 		updatechar(currchar[s]);
 	}
 	breakConcentration( currchar[s], s );
-	if( chars[currchar[s]].spellCast && ( chars[currchar[s]].casting == -1 || chars[currchar[s]].casting == 1 ) )
+	if( chars[currchar[s]].spellCast && ( chars[currchar[s]].casting == -1 || chars[currchar[s]].casting == 1))
 	{
 		sysmessage( s, "You can't do that while you are casting" );
 		return;
@@ -4114,57 +4114,57 @@ void cSkills::SkillUse(int s, int x) // Skill is clicked on the skill list
 	{
    case ARMSLORE:
 	   target(s, 0, 1, 0, 29, "What item do you wish to get information about?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case ANATOMY:
 	   target(s, 0, 1, 0, 37, "Whom shall I examine?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case ITEMID:
 	   target(s, 0, 1, 0, 40, "What do you wish to appraise and identify?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case EVALUATINGINTEL:
 	   target(s, 0, 1, 0, 41, "What would you like to evaluate?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case TAMING:
 	   target(s, 0, 1, 0, 42, "Tame which animal?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case HIDING:
 	   Skills->Hide(s);
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime+ (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case STEALTH:
 	   Skills->Stealth(s);
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case DETECTINGHIDDEN:
 	   target(s, 0, 1, 0, 77, "Where do you wish to search for hidden characters?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case PEACEMAKING:
 	   Skills->PeaceMaking(s);
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case PROVOCATION:
 	   target(s, 0, 1, 0, 79, "Whom do you wish to incite?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case ENTICEMENT:
 	   target(s, 0, 1, 0, 81, "Whom do you wish to entice?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case SPIRITSPEAK:
 	   Skills->SpiritSpeak(s);
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case STEALING:
 	   if (server_data.rogue)
 	   {
 		   target(s,0,1,0,205, "What do you wish to steal?");
-		   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+		   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 		   return;
 	   } else {
 		   sysmessage(s, "Contact your shard operator if you want stealing available.");
@@ -4172,33 +4172,33 @@ void cSkills::SkillUse(int s, int x) // Skill is clicked on the skill list
 	   }
    case INSCRIPTION:
 	   target(s, 0, 1, 0, 160, "What do you wish to place a spell on?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case TRACKING:
 	   Skills->TrackingMenu(s,TRACKINGMENUOFFSET);
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case BEGGING:
 	   target(s, 0, 1, 0, 152, "Whom do you wish to annoy?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case ANIMALLORE:
 	   target(s, 0, 1, 0, 153, "What animal do you wish to get information about?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case FORENSICS:
 	   target(s, 0, 1, 0, 154, "What corpse do you want to examine?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case POISONING:
 	   target(s, 0, 1, 0, 155, "What poison do you want to apply?");
-	   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+	   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   return;
    case MEDITATION:  //Morrolan - Meditation
 	   if(server_data.armoraffectmana)
 	   {
 		   Skills->Meditation(s);
-		   chars[currchar[s]].skilldelay=uiCurrentTime+(server_data.skilldelay*CLOCKS_PER_SEC);
+		   chars[currchar[s]].skilldelay = uiCurrentTime + (server_data.skilldelay * CLOCKS_PER_SEC);
 	   }
 	   else sysmessage(s, "Meditation is turned off.  Tell your GM to enable ARMOR_AFFECT_MANA_REGEN in server.scp to enable it.");
 	   return;
@@ -4216,7 +4216,7 @@ void cSkills::RandomSteal(int s)
 	char temp2[512];
 	tile_st tile;
 	
-	serial=calcserial(buffer[s][7],buffer[s][8],buffer[s][9],buffer[s][10]);
+	serial = calcserial(buffer[s][7],buffer[s][8],buffer[s][9],buffer[s][10]);
 	npc=findbyserial(&charsp[serial%HASHMAX], serial,1);
 	if (npc==-1) return;
 	p=packitem(npc);
@@ -4473,8 +4473,8 @@ void cSkills::Tracking(int s,int selection)
 {
 	int i = currchar[s];
 	chars[i].trackingtarget=chars[i].trackingtargets[selection]; // sets trackingtarget that was selected in the gump
-	chars[i].trackingtimer=((((tracking_data.basetimer*chars[i].skill[TRACKING])/1000)+1)*CLOCKS_PER_SEC)+uiCurrentTime; // tracking time in seconds ... gm tracker -> basetimer + 1 seconds, 0 tracking -> 1 sec, new calc by LB
-	chars[i].trackingdisplaytimer = tracking_data.redisplaytime * CLOCKS_PER_SEC + uiCurrentTime;
+	chars[i].trackingtimer = ((((tracking_data.basetimer*chars[i].skill[TRACKING])/1000)+1)*CLOCKS_PER_SEC) + uiCurrentTime; // tracking time in seconds ... gm tracker -> basetimer + 1 seconds, 0 tracking -> 1 sec, new calc by LB
+	chars[i].trackingdisplaytimer = (unsigned int) (tracking_data.redisplaytime * CLOCKS_PER_SEC + uiCurrentTime);
 	sprintf(temp,"You are now tracking %s.",chars[chars[i].trackingtarget].name);
 	sysmessage(s,temp);
 	Skills->Track(i);
